scripts/requirements-convert.txt
beautifulsoup4>=4.12.3
lxml>=5.2.2
markdownify>=0.13.1
PyYAML>=6.0.2
markitdown>=0.0.1a3
